requirements.txt
Django==5.1.3
pillow
gunicorn
whitenoise
python-decouple
django-cloudinary-storage